refactor(logic): centralize fog, topology, worker economy, AI control
- Extract shared fog-of-war and spy visibility helpers and reuse them from end-of-turn, combat, and quick battle resolution.
- Introduce MapTopology and connectivity helpers for province/sea adjacency and apply them in movement and connectivity_resolver.
- Move worker labour and economy primitives into a dedicated module and decouple AI control checks into ai_control while keeping ai_planner thin.

Made-with: Cursor
