+++ packages/colonizethis_logic/lib/src/turn/end_of_turn_resolver.dart
@@ -6,7 +6,7 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:logger/logger.dart';
 
 import '../dossier/event_dialogue.dart';
-import '../world/player_view.dart';
+import '../world/fog_resolution.dart';
 
 final Logger _log = Logger();
 
@@ -63,117 +63,6 @@ void _emitEraChangeDialogue(
   final events = dialogueEventsForEraChange(game, previousEra, newEra, seed);
   for (final e in events) onDialogue(e);
 }
-
-/// Spy 5-turn fog decay: decrement timers; when they expire, set other-faction
-/// provinces back to fogged for that player. Timers MUST NOT affect a player's
-/// own provinces; own provinces remain fully visible. SPEC/program/fog-and-exploration-resolution.md.
-(Map<String, Map<String, String>>, Map<String, Map<String, int>>)
-    applySpyRevealTimerDecay(Game game) {
-  final world = game.worldState;
-  final tileKeysByRegion = world.tileKeysByRegionAndProvince;
-  var visibilityByTile = Map<String, Map<String, String>>.from(
-    world.playerVisibilityByTile.map(
-      (k, v) => MapEntry(k, Map<String, String>.from(v)),
-    ),
-  );
-
-  // Province ownership lookup so we can ensure timers only affect other-faction provinces.
-  final ownerByProvinceId = <String, String?>{
-    for (final p in world.oldWorld.provinces) p.id: p.ownerId,
-    for (final p in world.newWorld.provinces) p.id: p.ownerId,
-  };
-
-  final nextSpyTimers = <String, Map<String, int>>{};
-  for (final entry in world.spyRevealTurnsByPlayer.entries) {
-    final playerId = entry.key;
-    final byProvince = entry.value;
-    final newByProvince = <String, int>{};
-    final vis = Map<String, String>.from(visibilityByTile[playerId] ?? {});
-    for (final provEntry in byProvince.entries) {
-      final provinceId = provEntry.key;
-      final turns = provEntry.value;
-
-      // Never apply Spy timers to a player's own provinces; clear any such timers without changing visibility.
-      final ownerId = ownerByProvinceId[provinceId];
-      if (ownerId == playerId) {
-        continue;
-      }
-
-      final nextTurns = turns - 1;
-      if (nextTurns <= 0) {
-        final regionId = ProvinceId.regionIdFrom(provinceId);
-        final tileKeys = tileKeysByRegion[regionId]?[provinceId] ?? [];
-        for (final tk in tileKeys) {
-          vis[tk] = VisibilityLevel.fogged.name;
-        }
-      } else {
-        newByProvince[provinceId] = nextTurns;
-      }
-    }
-    if (newByProvince.isNotEmpty) nextSpyTimers[playerId] = newByProvince;
-    visibilityByTile[playerId] = vis;
-  }
-  return (visibilityByTile, nextSpyTimers);
-}
-
-/// For each player, set tiles in other-faction provinces to fogged when no Explorer/Spy in that province.
-/// SPEC/program/fog-and-exploration-resolution.md.
-Map<String, Map<String, String>> applyFogDecay(Game game) {
-  const explorerTypes = {'explorer', 'spy'};
-  final owOwnerByProvince = {
-    for (final p in game.worldState.oldWorld.provinces) p.id: p.ownerId,
-  };
-  final nwOwnerByProvince = {
-    for (final p in game.worldState.newWorld.provinces) p.id: p.ownerId,
-  };
-
-  final provincesWithExplorerByPlayer = <String, Set<String>>{};
-  for (final u in game.worldState.oldWorld.units) {
-    if (explorerTypes.contains(u.type.toLowerCase())) {
-      provincesWithExplorerByPlayer
-          .putIfAbsent(u.ownerId, () => <String>{})
-          .add(u.locationProvinceId);
-    }
-  }
-  final provincesWithSpyTimerByPlayer = <String, Set<String>>{};
-  for (final entry in game.worldState.spyRevealTurnsByPlayer.entries) {
-    final playerId = entry.key;
-    final provinces = entry.value.keys;
-    if (provinces.isEmpty) continue;
-    provincesWithSpyTimerByPlayer[playerId] = provinces.toSet();
-  }
-  for (final u in game.worldState.newWorld.units) {
-    if (explorerTypes.contains(u.type.toLowerCase())) {
-      provincesWithExplorerByPlayer
-          .putIfAbsent(u.ownerId, () => <String>{})
-          .add(u.locationProvinceId);
-    }
-  }
-
-  final result = <String, Map<String, String>>{};
-  for (final entry in game.worldState.playerVisibilityByTile.entries) {
-    final playerId = entry.key;
-    final visibility = Map<String, String>.from(entry.value);
-    final hasExplorerIn = provincesWithExplorerByPlayer[playerId] ?? const {};
-    final hasSpyTimerIn = provincesWithSpyTimerByPlayer[playerId] ?? const {};
-
-    for (final tileKey in visibility.keys.toList()) {
-      final parts = tileKey.split('|');
-      if (parts.length != 4) continue;
-      final fullProvinceId = ProvinceId.full(parts[0], parts[1]);
-      final ownerId = owOwnerByProvince[fullProvinceId] ??
-          nwOwnerByProvince[fullProvinceId];
-      if (ownerId == null || ownerId == playerId) continue;
-      if (!hasExplorerIn.contains(fullProvinceId) &&
-          !hasSpyTimerIn.contains(fullProvinceId)) {
-        visibility[tileKey] = VisibilityLevel.fogged.name;
-      }
-    }
-    result[playerId] = visibility;
-  }
-  return result;
-}
-
 /// Returns the id of a Great Power that controls 31+ Old World provinces, or null.
 String? findMilitaryVictoryWinner(Game game) {
   const int requiredProvinces = 31;
